chore: decouple field menu components
