+++ packages/frontend/src/features/semanticViewer/components/SidebarViewFieldMenu/index.ts
@@ -1,0 +1,2 @@
+export { default as FieldFilterItems } from './FieldFilterItems';
+export { default as FieldTimeGranularityItems } from './FieldTimeGranularityItems';
